# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoulsGuide HTTP API",
    "version": "5.0.0",
    "description": "SoulsGuide 可执行路由契约。CI/后端测试会校验这里的 method/path 与 Express 实现双向一致。所有 JSON 响应使用 {code,message,data} 信封。"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署"
    }
  ],
  "paths": {
...
    "/api/media/upload/video/merge": {
      "post": {
        "summary": "校验、合并并转码视频",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": ["creator", "admin"],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "Success": {
        "description": "统一成功信封",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Envelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Envelope": {
        "type": "object",
        "required": ["code", "message", "data"],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
...
    }
  }
}